backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
pillow==10.1.0
opencv-python==4.8.1.78
scikit-image==0.22.0
google-cloud-vision==3.4.4
google-cloud-language==2.11.1
google-cloud-translate==3.11.1
google-cloud-monitoring==2.16.0
transformers==4.35.2
torch==2.1.0
spacy==3.7.2
beautifulsoup4==4.12.2
selenium==4.15.2
faker==20.1.0
firebase-admin==6.2.0
redis==5.0.1
celery==5.3.4
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0

